# 8361352: [lworld] Weird interaction between OSR and value class instances