--- a/src/hotspot/share/ci/ciEnv.hpp
+++ b/src/hotspot/share/ci/ciEnv.hpp
@@ -504,8 +504,8 @@ public:
   void dump_compile_data(outputStream* out);
   void dump_replay_data_version(outputStream* out);
 
-  ciWrapper* make_larval_wrapper(ciType* type) {
-    return _factory->make_larval_wrapper(type);
+  ciWrapper* make_early_larval_wrapper(ciType* type) {
+    return _factory->make_early_larval_wrapper(type);
   }
 
   ciWrapper* make_null_free_wrapper(ciType* type) {