--- a/src/hotspot/share/ci/ciObjectFactory.hpp
+++ b/src/hotspot/share/ci/ciObjectFactory.hpp
@@ -138,6 +138,7 @@ public:
 
   ciReturnAddress* get_return_address(int bci);
 
+  ciWrapper* make_larval_wrapper(ciType* type);
   ciWrapper* make_null_free_wrapper(ciType* type);
 
   GrowableArray<ciMetadata*>* get_ci_metadata() { return &_ci_metadata; }