--- a/src/hotspot/share/ci/ciType.cpp
+++ b/src/hotspot/share/ci/ciType.cpp
@@ -140,3 +140,16 @@ void ciReturnAddress::print_impl(outputStream* st) {
 ciReturnAddress* ciReturnAddress::make(int bci) {
   GUARDED_VM_ENTRY(return CURRENT_ENV->get_return_address(bci);)
 }
+
+ciWrapper::ciWrapper(ciType* type, int properties)
+    : ciType(type->basic_type()),
+      _type(type),
+      _properties(properties) {
+  assert(!type->is_wrapper(), "Thou shall not double wrap!");
+  assert(type->is_inlinetype()
+             // An abstract value type is an instance_klass
+             || (type->is_instance_klass() && type->as_instance_klass()->flags().is_abstract() && !type->as_instance_klass()->flags().is_identity())
+             // An unloaded inline type is an instance_klass (see ciEnv::get_klass_by_name_impl())
+             || (type->is_instance_klass() && !type->is_loaded()),
+         "should only be used for inline types");
+}