--- a/src/hotspot/share/ci/ciTypeFlow.hpp
+++ b/src/hotspot/share/ci/ciTypeFlow.hpp
@@ -853,7 +853,7 @@ public:
                                       return _block_map[rpo]; }
   int inc_next_pre_order()          { return _next_pre_order++; }
 
-  ciType* mark_as_larval(ciType* type);
+  ciType* mark_as_early_larval(ciType* type);
   ciType* mark_as_null_free(ciType* type);
 
 private: