--- a/src/hotspot/share/opto/parse1.cpp
+++ b/src/hotspot/share/opto/parse1.cpp
@@ -153,7 +153,7 @@ Node* Parse::fetch_interpreter_state(int index,
 // not a general type, but can only come from Type::get_typeflow_type.
 // The safepoint is a map which will feed an uncommon trap.
 Node* Parse::check_interpreter_type(Node* l, const Type* type,
-                                    SafePointNode* &bad_type_exit) {
+                                    SafePointNode* &bad_type_exit, bool is_larval) {
   const TypeOopPtr* tp = type->isa_oopptr();
 
   // TypeFlow may assert null-ness if a type appears unloaded.
@@ -184,10 +184,7 @@ Node* Parse::check_interpreter_type(Node* l, const Type* type,
       bad_type_exit->control()->add_req(bad_type_ctrl);
     }
 
-    // In an OSR compilation, we cannot know if a value object in the incoming state is larval or
-    // not. As a result, we must pass maybe_larval == true to not eagerly scalarize the result if
-    // the target type is a value class.
-    l = gen_checkcast(l, makecon(tp->as_klass_type()->cast_to_exactness(true)), &bad_type_ctrl, false, true);
+    l = gen_checkcast(l, makecon(tp->as_klass_type()->cast_to_exactness(true)), &bad_type_ctrl, false, is_larval);
     bad_type_exit->control()->add_req(bad_type_ctrl);
   }
 
@@ -378,7 +375,8 @@ void Parse::load_interpreter_state(Node* osr_buf) {
       // value and the expected type is a constant.
       continue;
     }
-    set_local(index, check_interpreter_type(l, type, bad_type_exit));
+    bool is_larval = osr_block->flow()->local_type_at(index)->is_larval();
+    set_local(index, check_interpreter_type(l, type, bad_type_exit, is_larval));
   }
 
   for (index = 0; index < sp(); index++) {
@@ -386,7 +384,8 @@ void Parse::load_interpreter_state(Node* osr_buf) {
     Node* l = stack(index);
     if (l->is_top())  continue;  // nothing here
     const Type *type = osr_block->stack_type_at(index);
-    set_stack(index, check_interpreter_type(l, type, bad_type_exit));
+    bool is_larval = osr_block->flow()->stack_type_at(index)->is_larval();
+    set_stack(index, check_interpreter_type(l, type, bad_type_exit, is_larval));
   }
 
   if (bad_type_exit->control()->req() > 1) {