… test/micro/org/openjdk/bench/valhalla/loops/osr/LarvalDetectionAboveOSR.java