@marc-chevalier 2025-08-26
After the successful testing with asserting that the abstract was useless (implied by instanceKlass without identity), I've push the fix. The condition is actually slightly relaxed as it also catches final value classes, and not only abstract value classes, but anyway, we want both.